Test config wrapper
Some refactoring is probably needed here.
If a configFile is passed to get, get will
read that file and cache it to avoid thrashing
the filesystem. I don't think it's normal behavior
to be passing in alternate configs during runtime.

In fact I'm declaring that unsupported right now.
$configFile parameter is really just for testing, but
whatever is passed to the first usage of Config::get
will be the file used for all subsequent calls.