Return jwt from user registration (#10)
JWTs (json web tokens) are lightweight authentication tokens.
This uses the firebase implementation of JWT to send back an
authentication token from register.

Set this token on the payload of any request which needs to be made by an authenticated user (any write other than login/registration) and then use JWT::decode($token, $key, array('HS256' /* in our case */)) and read the userId from that object to check that the user has been issued a JWT.

### security note
encryption key for creating the jet is currently hardcoded into the source. This is a terrible practice, and should never be done. This should really be updated to read from an environment variable, but key management is hard and this is a school project application. I think we're ok to not have to deal with the problems of key management.

* Add firebase/php-jwt as a dep
* Return a jwt from user registration
* Display errors in registation form